Servicio dietas + cambio PK en entidades a Long + cambios en repositorios
Se ha creado el servicio de las dietas, con sus excepciones correspondientes, así como modificar las entidades para que su clave primaria sea de tipo Long, y sus repositorios. Al repositorio de Dieta se añadió un findByNombre para el servicio.

+++ jpa/src/main/java/es/uma/sisinfparInt/dns/jpa/repositories/ClienteRepository.java
@@ -7,15 +7,15 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import es.uma.sisinfparInt.dns.jpa.entities.Cliente;
 
-public interface ClienteRepository extends JpaRepository<Cliente,Integer> {
-    Optional<Cliente> findById(Integer id); //Busca un cliente por una id
+public interface ClienteRepository extends JpaRepository<Cliente,Long> {
+    Optional<Cliente> findById(Long id); //Busca un cliente por una id
 
     List<Cliente> findAll(); //Busca todos los clientes
 
     Cliente save(Cliente cliente); //Guarda o actualiza un cliente
 
 
-    void deleteById(Integer id); //Borra un cliente segun la id
+    void deleteById(Long id); //Borra un cliente segun la id
 
     void deleteAll(); //Borra todos los clientes
 }
